Add call to action for suggestion readers
There's some discussion on Slack about how to instruct readers about interaction with the suggestion issues. The main README says to :+1: issues that you want to see implemented, but when issues are directly linked that guidance is never seen. This attempts to call out precisely what's expected of a reader.